Get rid of duplicate Converter code, and re-establish Version numbers for Simulations.
