+++ Models/Models.csproj
@@ -142,7 +142,6 @@
     <Compile Include="Core\APSIMFileWriter.cs" />
     <Compile Include="Core\Attributes\DoNotDocumentAttribute.cs" />
     <Compile Include="Core\AutoDocumentation.cs" />
-    <Compile Include="Core\Converter.cs" />
     <Compile Include="Core\Alias.cs" />
     <Compile Include="Core\Events.cs" />
     <Compile Include="Core\FileFormat.cs" />
